starting to be a game

+++ Assets/Scripts/ScriptableObjs/EnemySO.cs
@@ -5,9 +5,9 @@ public class EnemySO : ScriptableObject
 {
     public string enemyName;
     public Sprite sprite;
-    public int health = 100;
+    public float health = 100;
     public float moveSpeed = 1f;
     public float waitTime = 2.5f;
-    public int damage = 10;
+    public float damage = 10;
     public RuntimeAnimatorController animatorController;
 }
